Fix overflow-safe size calculations in depth clear and benchmark
Agent-Logs-Url: https://github.com/starseeker/osmesa/sessions/781d7227-9b79-48e7-9ac4-fc518061f23c

+++ examples/triangle_bench.c
@@ -199,7 +199,8 @@ main(int argc, char **argv)
 	return 1;
     }
 
-    framebuffer = (GLubyte *) calloc((size_t) width * height * 4, sizeof(GLubyte));
+    framebuffer = (GLubyte *) calloc((size_t) width * (size_t) height * 4,
+				     sizeof(GLubyte));
     if (!framebuffer) {
 	fprintf(stderr, "framebuffer allocation failed\n");
 	OSMesaDestroyContext(ctx);
