+++ src/swrast/s_depth.c
@@ -1325,7 +1325,7 @@ _swrast_clear_depth_buffer(GLcontext *ctx, struct gl_renderbuffer *rb)
 		if (contiguous) {
 		    /* optimized case */
 		    GLuint *dst = (GLuint *) rb->GetPointer(ctx, rb, x, y);
-		    size_t len = (size_t) width * height * sizeof(GLuint);
+		    size_t len = (size_t) width * (size_t) height * sizeof(GLuint);
 		    memset(dst, (clearValue & 0xff), len);
 		    optimized = GL_TRUE;
 		}
